2020-09-29 Commit

+++ src/com/animalfighter/api/service/IFightAnimal.java
@@ -3,7 +3,7 @@ package com.animalfighter.api.service;
 import com.animalfighter.entities.Animal;
 
 public interface IFightAnimal {
-    public int calculateStrength(int type, int weight, int strength);
+    public double calculateStrength(int type, int weight, double strength);
 
     public void fightAnimal(Animal animal1, Animal animal2);
 }
